dom/plugins/test/mochitest/test_enumerate.html

<html>
<head>
  <title>NPAPI Cookie Tests</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="utils.js"></script>

  <link rel="stylesheet" type="text/css" 
        href="/tests/SimpleTest/test.css" />
</head>

<body onload="runTests()">
  <script class="testbody" type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

    function runTests() {
      var pluginElement = document.getElementById("plugin1");
      var c = 0;
      var foundSetColor = false;
      for (var n in pluginElement) {
        ++c;
        ok(n in pluginElement, "Enumerated property which doesn't exist?");
        if (n == 'setColor')
          foundSetColor = true;
      }
      ok(c > 0, "Should have enumerated some properties");
      ok(foundSetColor, "Should have enumerated .setColor");
      SimpleTest.finish();
    }
  </script>

  <p id="display"></p>

  <embed id="plugin1" type="application/x-test" width="400" height="400"></embed>
</body>
</html>
